works but needs more config to be done

+++ requirements/wp/install_wp.sh
@@ -4,11 +4,18 @@
 set -e
 
 cd /var/www/html
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-chmod +x wp-cli.phar
-./wp-cli.phar core download --allow-root
-./wp-cli.phar config create --dbname=wordpress --dbuser=wpuser --dbpass=password --dbhost=mariadb --allow-root
-./wp-cli.phar core install --url=localhost --title=inception --admin_user=admin --admin_password=admin --admin_email=admin@admin.com --allow-root
+
+# Check if WordPress is already installed
+if [ ! -f wp-config.php ]; then
+  echo "Downloading and configuring WordPress..."
+  curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+  chmod +x wp-cli.phar
+  ./wp-cli.phar core download --allow-root
+  ./wp-cli.phar config create --dbname=wordpress --dbuser=wpuser --dbpass=password --dbhost=db --allow-root
+  ./wp-cli.phar core install --url=localhost --title=inception --admin_user=admin --admin_password=admin --admin_email=admin@admin.com --allow-root
+else
+  echo "WordPress is already installed."
+fi
 
 # Start php-fpm in the foreground
 php-fpm7.4 -F
